integrate unittest into test-mp-equation
